!Убрал позорный варнинг
git-svn-id: svn://svn.boba.su@1677 98664855-a6bc-5d49-be9f-2936a658fbf1

+++ ARM/XTASK3/gui.c
@@ -477,7 +477,7 @@ void bm_menu_iconhndl(void *data, int curitem, void *unk)
           char *p=strrchr(s,'\\');
           if (p)
           {
-            short pos;
+            unsigned int pos;
             ws=AllocWS((l=strlen(p+1)));
             str_2ws(ws,p+1,l);
             if ((pos=wstrrchr(ws,ws->wsbody[0],'.'))!=0xFFFF)
